feat(mod): wire permissions + faction binding + fog-of-war (Phase 3)
Per-agent permission gates and PvP fog-of-war scoping. Every action / control /
read command now declares the permission it needs via ctx.Require(...) at the
top of ExecuteAsync; the bridge maps the resulting BridgeRejectionException
to HTTP 403 with structured error code (PERMISSION_DENIED /
FACTION_SCOPE_VIOLATION).

Permission matrix wired:
- Read (list_kingdoms, list_cities, query_actors)  : ReadOwnFaction
- Screenshot                                       : ReadAll  (factionplayers
                                                     can't snap to bypass fog)
- Action (invoke_power, spawn)                     : ActionFaction | ActionGlobal
- Paint tile                                       : ActionGlobal  (terraforming
                                                     is genuinely god-only)
- Control (pause/resume/speed/generate/save/load)  : ControlWorld

Fog-of-war filter applied to list_cities + query_actors: cities/actors of
kingdoms the agent can't see are dropped from the result. list_kingdoms is
intentionally NOT filtered — agents need to know enemy factions exist for
diplomacy / threat awareness. ReadAll holders (god, observer, narrator) and
sessions with partial_intel=false always see everything.

Refactors enabling Phase 3:
- BridgeRejectionException moved out of InvokePowerCommand.cs into its own
  file so the test project can link it without Unity references.
- ErrorCode constants moved out of ErrorEnvelope.cs into ErrorCode.cs (no
  Newtonsoft.Json dep) so they're linkable too.
- Agent.ClaimedKingdomId migrated from int? to long? — kingdom IDs are long
  in the game, the int? was a leftover from early sketching.

New RequestContext helpers (all covered by RequestContextTests):
- HasAny / RequireAny — for OR-gates like "ActionFaction or ActionGlobal".
- CanSeeKingdom(long) — predicate for read-side fog-of-war filtering.
- RequireKingdomAccess(long) — strict gate for action commands that target a
  specific kingdom.

Tests: 22 new xunit cases (47 total, all green) covering permission flags per
role, RequireAny/Require behavior, CanSeeKingdom matrix (7 InlineData rows
across role × claim × partial_intel × target), RequireKingdomAccess
(god / unbound / own / foreign kingdom paths).

Live-validated against a running game with the 3-agent PvP agents.json:
12/12 cases pass (4 athena reads OK, 5 athena denials with PERMISSION_DENIED,
narrator reads OK, narrator invoke_power denial, narrator screenshot OK).

+++ mod/src/WorldBoxBridge/Session/Agent.cs
@@ -4,7 +4,7 @@ namespace WorldBoxBridge.Session;
 
 public sealed class Agent
 {
-    public Agent(string id, string token, AgentRole role, int? claimedKingdomId, Permission permissions)
+    public Agent(string id, string token, AgentRole role, long? claimedKingdomId, Permission permissions)
     {
         Id = id ?? throw new ArgumentNullException(nameof(id));
         Token = token ?? throw new ArgumentNullException(nameof(token));
@@ -16,7 +16,7 @@ public sealed class Agent
     public string Id { get; }
     public string Token { get; }
     public AgentRole Role { get; }
-    public int? ClaimedKingdomId { get; set; }
+    public long? ClaimedKingdomId { get; set; }
     public Permission Permissions { get; }
     public DateTime LastSeenUtc { get; set; }
 }
